requirements.txt
aniso8601==4.0.1
certifi==2018.10.15
cffi==1.11.5
Click==7.0
Flask==1.0.2
Flask-Sockets==0.2.1
gevent==1.3.7
gevent-websocket==0.10.1
greenlet==0.4.15
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.2.5
MarkupSafe==1.1.0
pycparser==2.19
python-docx==0.8.7
pytz==2018.7
six==1.11.0
websocket-client==0.54.0
Werkzeug==0.14.1
wincertstore==0.2
